suzerain: fix signed/unsigned warning
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44548 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/reacting/main_advance.cpp
@@ -514,7 +514,7 @@ suzerain::reacting::driver_advance::run(int argc, char **argv)
 
         // growth rate amplitude for species must be set all equal for consistency
         // setting them all equal to zero
-        for (int s=0; s<Ns-1; s++) {
+        for (size_t s=0; s<Ns-1; s++) {
             sgdef->gramp_mean[is0+s] = sgdef->gramp_mean[0];
         }
